mark web and common files with TODO headers
The items in util/hooks assume constant internet connectivity and it
will only cache data once per session, because the data is cached in
TypeScript variables. That is, refreshing the page on  web or (force)
closing and reopening the app on mobile will clear the cache. This
isn't proper for mobile, because we can't always assume that mobile has
internet connectivity.