CI: run every non-deploy job if CI YAML changes
The rationale here is that if any changes are made to the CI
configuration, every CI job should be verified to ensure they still all
still work. There's no way to verify what individual job configurations
have changed within the YAML file, so this is the best we can do. This
also lets us make MRs that only change the YAML and nothing else.