add Mocha + Chai tests for common
This moves variables in auth.ts in order to let the test code call
shared functions while still ensuring that they are not exported for use
outside of common.

For CI, we added tests with test reports and also a compiler check (via
`npx tsc`) for common to ensure that common actually compiles correctly.