handle non-JSON responses from failed HTTP requests
The `.json()` call might fail if the page returns some html page as the
response and not a JSON, resulting in the parsing error being used
instead. As this is less descriptive, support undefined responses in
APIFetchFailError so that callers can use the HTTP error status code as
the human-readable error message.