client: for pictures, require auth and standardize file names
The previous iteration involved using `Math.random()` as the main
security feature to protect access to client images. However,
`Math.random()` isn't cryptographically secure; the documentation for it
also says to not use it for anything related to security. This
additionally results in more orphaned images whenever an edit is made,
as the server doesn't have a cleanup function for orphaned images right
now, so there will just be a bunch of image files with random strings as
the name.

Fix this (for images going forward) by requiring authentication (auth
header) to get a client's picture and disallowing access to the raw
uploads directory. We also now just name the client pictures by the
client's autoincrementing primary key (id). Doing this allows for the
previous client picture to be overwritten. (There's not much of a use
case to keep a history of client pictures on the backend right now.)