client: add modified timestamp and picture cache headers
The previous implementation served images directly from the web
server (Caddy in our case), which allowed it to automatically to set
caching headers. However, Django responses don't do this by default. Set
up the backend so that it sends ETag and Last-Modified headers for
client pictures to allow clients to cache the responses appropriately.