fix new client pictures saving as "client-None.png"
The client picture files are named using the primary autoincrementing
keys, but a new client will not have this key set yet until it is saved
into the database. Doing this is problematic, since creating new clients
after that will overwrite the image and cause two clients to have the
same image. Re-save image files for new clients to fix this issue,
making sure to clean up the orphaned temporary files.