optimize PhotoViewUpload by not using data URLs
Data URLs are not optimal and cause a 33% increase in size (they're
base64-encoded strings of photo data). The `toDataURL` calls are also
synchronous (blocks the UI), eats memory, and is overall slower to work
with [1].

Optimize the image cropper by not using data URLs and working with blobs
directly. Revoke the blobs when they're being replaced in order to free
memory. This is better than worrying about the Formik values prop or the
cropper holding a reference to the base64 image data, since that
prevents garbage collection. The downside to this optimization is that
everything starts to feel like C again!

[1] https://stackoverflow.com/a/59025746