fix wrong image on client picture edit discard
A bug can occur if these steps are followed:

1. Open a client page and note the initial picture state of the client.
2. Edit the client, pick image X, crop it, save it in the cropper, then
   save the form (make the PUT request) to use image X as the new
   picture for the client.
3. In the same session (no refresh), repeat step 2, but do not submit
   the form. Instead, after saving the crop, press cancel.

This bug results in the initial picture state from step 1 being used.
Fix this by detecting whether an image has ever been loaded before
attempting a GET request for the client photo. This results in definite
network requests when pressing cancel after saving a crop but not
submitting the form; however, the client image itself is cached, so it's
unlikely a full download will occur.

We also ensure the current picture has its blob released when the
PhotoViewUpload component is unmounted.