handle dashboard errors better
If an exception occurs, all the `.catch` handler does is make the
variable undefined, which causes a promise rejection in `fetchClients`
and `fetchReferrals` since the code tries to access properties of
undefined objects. This causes an entire error screen to show up on
development builds because of the unhandled promise rejections. Telling
the user that "something went wrong" is also vague.

Fix this by showing an alert component with the exception's error
message.