web: fix admin and stats pages not loading
The admin and stats page depends on the current user. There is code
somewhere on web that sends the initial request to get the current user.

The cache hook initially use `_value` and only calls
`getCachedValueInner()` if `isInvalidated` is true. However, on app
startup, `isInvalidated` is false because there is an initial promise.
This means the hook won't use the initial promise's value when it is
resolved and the hook continues to use the undefined loading value.

Fix this by always calling `getCachedValueInner` in the hook initially,
as that will use the initial promise if it exists. This should not cause
a re-render if `_value` passes `Object.is` comparison with the promise's
value [1].

[1] https://reactjs.org/docs/hooks-reference.html#bailing-out-of-a-state-update