use TextInput as anchor for dropdown menu
The previous implementation used an invisible anchor in an attempt to
get the menu to display right below the TextInput component like an
extension. However, this was too hacky, resulting in issues such as the
gap between the top of the menu and the TextInput being different across
multiple screens. This is suspected to be how React Native Paper is
rendering the Portal components, which is what the Menu component from
Paper uses.

We stop using a nearly invisible anchor (i.e., give up on trying to put
the menu right below the TextInput) and instead opt to display the
selection as a highlighted menu item. This results in the menu being
often being displayed directly above the TextInput, which is similar to
how the dropdown menu components on web perform (although mobile doesn't
align the current selection against the menu item of the selection).