memoize FormikTextInput to improve form performance
Formik apparently re-renders all components that use the FormikProps
when it is set to validate on every change (true by default). This
causes massive latency in Formik forms, especially when switching
between multiple FormikTextInputs. Even on production builds with the
standalone APK, the latency is noticeable.

Fix this by memoizing the FormikTextInput so that the TextInput and
HelperText are only re-rendered when the relevant properties change.
This significantly increases performance on Formik forms using
FormikTextInput.