ExposedDropdownMenu: improve `values` prop check
Certain places might pass in values created with `Object.entries` or
`Array.reduce` or other functions that create new objects everytime,
even if the contents of that new object are the same as the old one. So
we stop doing a reference check (i.e. stop using !==) on the `values`
prop and instead do a linear-scan compare of the contents. (Same as why
you would not want to do == to check for equality of two Collections in
Java, since == just checks the pointer value in Java.)

We also extend the memoization equality function of
FormikExposedDropdownMenu to include the values check.